requirements.txt
PyQt5>=5.15.0
mss>=9.0.0
pytesseract>=0.3.10
Pillow>=10.0.0
numpy>=1.24.0
opencv-python>=4.8.0
